rewrite `dict` node representation to reduce indirection
with this change, the struct `value-node` is completely eliminated; `hash-node`,
`conflict-node` or one-element `dict` simply hold the key and the value.

`conflict-node` is now a newtype around `simple-vector` which holds a series of key/value
pairs inline like a plist (except a vector). conflict-hashes are now stored by the parent,
rather than within the conflict-node. infrastructure is in place to add a header to the
front of a `conflict-node` to store e.g. a transient id, though currently
`conflict-node`'s header is empty.

`hash-node` is now a newtype around `simple-vector` which holds a header of
child-key/child-value pairs inline like a plist (except a vector). child-present-p bitmaps
are now stored by the parent, rather than within the hash-node. each of the contained
pairs may be:
- a key and a value.
- a conflict-hash and a conflict-node.
- a child-hash-node-contents bitmap and a hash-node.
`hash-node` has a header before its contents, which currently holds two bitmaps,
`child-is-conflict-p` and `child-is-entry-p`, which determine the type and meaning of the
child-key/child-value pairs.