halfway-between dict repr with somewhat reduced indirection
With this commit, `dict` nodes have three variants:
- `entry-node`, a struct which is effectively a key/value pair.
- `conflict-node`, a simple-vector which stores a header containing a conflict-hash,
  followed by some number of `entry-node`s.
- `hash-node`, a simple-vector which stores a header containing a bitmap, followed by some
  number of child `node`s.

With luck, this will be a pleasant middle ground between the naive repr, which had hash-
and conflict-nodes be structs containing simple-vectors; and the inline repr, which
eliminated `entry-node` and had conflict- and hash-nodes be simple-vectors containing
keys and values inline. The former had unnecessary redirection, while the latter involved
too much copying to be efficient.